k8s: add egressGateway property to CiliumEgressNATPolicy
This commit adds a new egressGateway property to the
CiliumEgressNATPolicy object, which improves the way the egress gateway
node is selected and configured.

Before this change a CiliumEgressNATPolicy would only allow to specify a
an egressSourceIP, which was used both as:

* IP to masquerade traffic leaving the cluster
* IP to identify which node was acting as gateway for the policy

Now, with the egressGateway property it's possible to explicitly
identify which node should act as gateway by using a nodeSelector based
on node labels. In this way, workers node will forward traffic to the
egress gateway node by using its internal node IP.

Moreover, the new property extends the way the egress IP can be
configured, as a user can either:

* specify the egress IP that should be used to masquerade traffic
* specify the interface that should be used to masquerade traffic: in
  this case the first IPv4 assigned to that interface will be used as
  egress IP
* not specify anything: in this case the first IPv4 assigned to the
  interface for the default route will be used

Signed-off-by: Gilberto Bertin <jibi@cilium.io>